Reorganise booking E2E tests
We reorganise our booking E2E tests, ahead of expanding them to cover editing departures and cancellations. To prevent essentially n-squared growth in booking test steps, we reorganise our booking tests, so we have a scenario per possible end state, rather than a scenario per possible intermediate and end state. We also examine the booking history page within the happy-path scenarios, rather than in its own scenarios. And we go back to the booking page after encountering an error when updating its state, ready for if we want to make another attempt at updating the booking
