+++ e2e/tests/stepDefinitions/manage/arrival.ts
@@ -55,13 +55,15 @@ Then('I should see the booking with the arrived status', () => {
 
     const bedspaceShowPage = Page.verifyOnPage(BedspaceShowPage, this.premises, this.room)
     bedspaceShowPage.shouldShowBookingDetails(this.booking)
+    bedspaceShowPage.clickBookingLink(this.booking)
   })
 })
 
 Then('I should see a list of the problems encountered marking the booking as arrived', () => {
   cy.then(function _() {
     const page = Page.verifyOnPage(BookingArrivalNewPage, this.premises, this.room, this.booking)
-
     page.shouldShowErrorMessagesForFields(['arrivalDate', 'expectedDepartureDate'])
+
+    page.clickBack()
   })
 })
