+++ e2e/tests/stepDefinitions/manage/booking.ts
@@ -2,6 +2,7 @@ import { Given, Then } from '@badeball/cypress-cucumber-preprocessor'
 import Page from '../../../../cypress_shared/pages/page'
 import BedspaceShowPage from '../../../../cypress_shared/pages/temporary-accommodation/manage/bedspaceShow'
 import BookingConfirmPage from '../../../../cypress_shared/pages/temporary-accommodation/manage/bookingConfirm'
+import BookingHistoryPage from '../../../../cypress_shared/pages/temporary-accommodation/manage/bookingHistory'
 import BookingNewPage from '../../../../cypress_shared/pages/temporary-accommodation/manage/bookingNew'
 import BookingShowPage from '../../../../cypress_shared/pages/temporary-accommodation/manage/bookingShow'
 import bookingFactory from '../../../../server/testutils/factories/booking'
@@ -75,6 +76,25 @@ Then('I should see a confirmation for my new booking', () => {
 
     const bedspaceShowPage = Page.verifyOnPage(BedspaceShowPage, this.premises, this.room)
     bedspaceShowPage.shouldShowBookingDetails(this.booking)
+    bedspaceShowPage.clickBookingLink(this.booking)
+  })
+})
+
+Then('I should see previous booking states in the booking history', () => {
+  cy.then(function _() {
+    const bookingShowPage = Page.verifyOnPage(BookingShowPage, this.premises, this.room, this.booking)
+    bookingShowPage.clickHistoryLink()
+
+    const bookingHistoryPage = Page.verifyOnPage(
+      BookingHistoryPage,
+      this.premises,
+      this.room,
+      this.booking,
+      this.historicBookings,
+    )
+    bookingHistoryPage.shouldShowBookingHistory()
+
+    bookingHistoryPage.clickBack()
   })
 })
