+++ e2e/tests/stepDefinitions/manage/cancellation.ts
@@ -63,13 +63,15 @@ Then('I should see the booking with the cancelled status', () => {
 
     const bedspaceShowPage = Page.verifyOnPage(BedspaceShowPage, this.premises, this.room)
     bedspaceShowPage.shouldShowBookingDetails(this.booking)
+    bedspaceShowPage.clickBookingLink(this.booking)
   })
 })
 
 Then('I should see a list of the problems encountered cancelling the booking', () => {
   cy.then(function _() {
     const page = Page.verifyOnPage(BookingCancellationNewPage, this.premises, this.room, this.booking)
-
     page.shouldShowErrorMessagesForFields(['date'], 'bookingCancellation')
+
+    page.clickBack()
   })
 })
