+++ e2e/tests/stepDefinitions/manage/departure.ts
@@ -59,13 +59,15 @@ Then('I should see the booking with the departed status', () => {
 
     const bedspaceShowPage = Page.verifyOnPage(BedspaceShowPage, this.premises, this.room)
     bedspaceShowPage.shouldShowBookingDetails(this.booking)
+    bedspaceShowPage.clickBookingLink(this.booking)
   })
 })
 
 Then('I should see a list of the problems encountered marking the booking as departed', () => {
   cy.then(function _() {
     const page = Page.verifyOnPage(BookingDepartureNewPage, this.premises, this.room, this.booking)
-
     page.shouldShowErrorMessagesForFields(['dateTime', 'reasonId', 'moveOnCategoryId'])
+
+    page.clickBack()
   })
 })
